backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==2.1.1
sentence-transformers==3.1.0
scikit-learn==1.5.1
python-multipart==0.0.9
google-generativeai==0.7.2
